requirements.txt
pandas==2.0.3
scikit-learn==1.3.0
python-owasp-zap-v2.4==0.0.22
numpy==1.24.3
